requirements.txt
fastapi
uvicorn
python-multipart
tencentcloud-sdk-python
